! 2009-06-11 - Chris Jones  - Fix WIN32 warnings about double -> float conversions
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@11064 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: d79b656ea048d8627c47fdac685de37457e9941e

+++ Rec/Rich/RichRecMCTools/src/RichPhotonCreatorWithGaussianCKSmear.cpp
@@ -4,7 +4,7 @@
  *  Implementation file for RICH reconstruction tool : RichPhotonCreatorWithGaussianCKSmear
  *
  *  CVS Log :-
- *  $Id: RichPhotonCreatorWithGaussianCKSmear.cpp,v 1.7 2007-02-02 10:06:27 jonrob Exp $
+ *  $Id: RichPhotonCreatorWithGaussianCKSmear.cpp,v 1.8 2009-06-11 12:41:07 jonrob Exp $
  *
  *  @author Chris Jones   Christopher.Rob.Jones@cern.ch
  *  @date   08/07/2004
@@ -163,7 +163,7 @@ PhotonCreatorWithGaussianCKSmear::buildPhoton( LHCb::RichRecSegment * segment,
                     << ": theta = " << newPhoton->geomPhoton().CherenkovTheta()
                     << " -> " << newCKtheta << endreq;
         }
-        newPhoton->geomPhoton().setCherenkovTheta( newCKtheta );
+        newPhoton->geomPhoton().setCherenkovTheta( static_cast<float>(newCKtheta) );
 
         // count smeared photons
         ++m_smearCount[rad];
